add telegram notifications

+++ src/utils.py
@@ -2,31 +2,32 @@ import os
 import cv2
 import datetime
 
+
 class Utils:
 
     @staticmethod
-    def save_capture(frame, class_name, path):
+    def save_capture(frame, path):
         """
         Guarda una imagen con timestamp en la carpeta de capturas.
 
         :param frame: Imagen a capturar
         :type frame: np.array
 
-        :param class_name: Nombre del objeto detectado
-        :type class_name: str
-
         :param path: Ruta para guardar la captura
         :type path: str
 
-        :return: None
-        :rtype: None
+        :return filepath: Ruta de la imagen guardada
+        :rtype filepath: str
         """
         os.makedirs(path, exist_ok=True)
 
-        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M")
-        filename = f"{class_name}_{timestamp}.jpg"
+        timestamp = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+        filename = f"{timestamp}.jpg"
 
         filepath = os.path.join(path, filename)
-        cv2.imwrite(filepath, frame)
+        if not os.path.exists(filepath):
+            cv2.imwrite(filepath, frame)
 
         print(f"Captura guardada como '{path}/{filename}'.")
+
+        return filepath
